Refactor package structure
Fix compilation errors due to package restructuring
