+++ core/src/main/java/com/example/demo/core/shoe/InMemoryShoeRepo.java
@@ -1,7 +1,6 @@
-package com.example.demo.core;
+package com.example.demo.core.shoe;
 
 import com.example.demo.dto.in.ShoeFilter;
-import com.example.demo.dto.in.ShoeToStock;
 
 import java.util.HashMap;
 import java.util.List;
