+++ core/src/main/java/com/example/demo/core/stock/InvalidStockQuantity.java
@@ -1,4 +1,4 @@
-package com.example.demo.core;
+package com.example.demo.core.stock;
 
 public class InvalidStockQuantity extends RuntimeException {
