+++ core/src/main/java/com/example/demo/core/stock/StockCore.java
@@ -1,4 +1,4 @@
-package com.example.demo.core;
+package com.example.demo.core.stock;
 
 import com.example.demo.dto.in.ShoeFilter;
 import com.example.demo.dto.in.ShoeToStock;
