+++ core/src/main/java/com/example/demo/core/stock/StockFacade.java
@@ -1,9 +1,6 @@
-package com.example.demo.core;
+package com.example.demo.core.stock;
 
-import com.example.demo.dto.in.ShoeFilter;
-import com.example.demo.dto.out.Shoes;
-import com.example.demo.dto.out.Stock;
-import com.example.demo.dto.out.Stock.StockState;
+import com.example.demo.core.stock.StockCore;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
